feat(snapshot): refactor holdings projection state
