+++ crates/core/src/portfolio/snapshot/holdings_calculator/handlers/mod.rs
@@ -1,0 +1,5 @@
+//! Per-activity handlers grouped by domain.
+mod cash_flows;
+mod corporate_actions;
+mod trades;
+mod transfers;
